Improve the package API.
Events is now a method returning the event channel. This way it can
implement an interface. Also, the, now unexported field of Tracee can
be a send/receive channel, simplifying some of the code.

Errors is now just Error, and it's a method not a channel. It is to be
called once after all events have been received to see if there was an
error while tracing.