Fix a race condition.
There was a race condition between methods sending on the Tracee.cmds
channel and the closing of the channel. Now, we use an rwlock to
synchronize the sends (which are reads, curiously enough) with the
close (which is a write). Methods that send on the command channel now
do so through the Tracee.do method, which encapsulates the use of the
rwlock and allows for graceful handling of the case where the tracee
exits (and the command channel is closed) before a command is
executed.